SS-Mamba/requirements.txt
# Core dependencies
torch>=2.0.0
torchvision>=0.15.0
numpy>=1.21.0
pandas>=1.3.0
scikit-learn>=1.0.0
scipy>=1.7.0

# Mamba dependencies
mamba-ssm>=1.0.0
causal-conv1d>=1.0.0
einops>=0.6.0

# WSI processing
openslide-python>=1.2.0
opencv-python>=4.5.0
Pillow>=9.0.0
h5py>=3.0.0

# Visualization
matplotlib>=3.4.0
seaborn>=0.11.0
tqdm>=4.60.0

# Configuration
pyyaml>=6.0
omegaconf>=2.1.0

# Logging
tensorboard>=2.8.0
wandb>=0.12.0

# Utilities
timm>=0.6.0
albumentations>=1.0.0
